patch: refactor, code style and break out classes to their own file
