+++ src/Providers/ReferenceNumbers/IReferenceNumberProvider.cs
@@ -1,0 +1,7 @@
+namespace form_builder.Providers.ReferenceNumbers
+{
+    public interface IReferenceNumberProvider
+    {
+        string GetReference(string prefix, int length=8);
+    }
+}
